+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 This version of `cargo-ndk` is much closer to a thin passthrough to `cargo` than previous versions. Please take note of the breaking changes.
 
+It also introduces a new command, `cargo ndk-test`. This will push your tests to a device via adb and run them on-device. It currently has 
+a limitation that doctests can still only be run on the host.
+
 **New MSRV: 1.86**
 
 - **Breaking change**: Replaced gumdrop CLI parsing with clap. This is functionally equivalent but may cause some minor behavioral differences in edge cases
@@ -11,7 +14,8 @@ This version of `cargo-ndk` is much closer to a thin passthrough to `cargo` than
 - **Breaking change**: Removed Cargo.toml-based configuration property handling
 - **Breaking change**: Bump minimum supported Rust version (MSRV) to 1.86 and use Rust edition 2024.
 - **Breaking change**: `--bindgen` has been removed, and relevant environment variables are set by default
-- Enhancement: Clang builtins can now be linked automatically by adding `--link-builtins` or its equivalent environment variable
+- Feature: Clang builtins can now be linked automatically by adding `--link-builtins` or its equivalent environment variable
+- Feature: `cargo ndk-test` for running tests on Android devices.
 - Enhancement: Can now be built for an Android host (i.e. Termux) if built with `CARGO_NDK_ON_ANDROID` environment variable set
 - Enhancement: CLI flags can now be used in any order (e.g., `cargo ndk -t x86 build` and `cargo ndk build --target x86` are equivalent).
 - Enhancement: Added environment variable support for configuration flags with `CARGO_NDK_` prefix (e.g., `CARGO_NDK_TARGET`, `CARGO_NDK_PLATFORM`, `CARGO_NDK_OUTPUT_DIR`)